requirements.txt
fastapi
tensorflow-lite
librosa
soundfile
pydub
uvicorn
